fix schema
fix schema descriotion
also upgrade footer position

+++ frontend/src/qraphql/mutations/categories.mutations.tsx
@@ -18,8 +18,7 @@ export const UPDATE_CATEGORY =
     }
   }
 `);
-export const DELETE_CATEGORY =
-  gql(`mutation DeleteCategory($categoryId: Float!) {
+export const DELETE_CATEGORY = gql(`mutation DeleteCategory($categoryId: Int!) {
     deleteCategory(categoryId: $categoryId)
   }
 `);
