+++ frontend/src/qraphql/queries/tasks.queries.tsx
@@ -1,6 +1,6 @@
 import { gql } from '../../__generated__';
 
-export const GET_ALL_TASKS = gql(`query AllTasks($categoryId: Float!) {
+export const GET_ALL_TASKS = gql(`query AllTasks($categoryId: Int!) {
     allTasks (categoryId: $categoryId){
         id
         name
@@ -12,7 +12,7 @@ export const GET_ALL_TASKS = gql(`query AllTasks($categoryId: Float!) {
 `);
 
 export const GET_ONE_TASK =
-  gql(`query TaskById ($categoryId: Float!,$taskId: Float!){
+  gql(`query TaskById ($categoryId: Int!,$taskId: Int!){
     taskById (categoryId: $categoryId, taskId: $taskId){
         id
         name
